move range cache code to new file
